requirements.txt
alembic==1.17.2
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
coverage==7.12.0
curl_cffi==0.13.0
cycler==0.12.1
fonttools==4.60.1
frozendict==2.4.7
idna==3.11
iniconfig==2.3.0
jsonlines==4.0.0
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
multitasking==0.0.12
numpy==2.3.4
packaging==25.0
pandas==2.3.3
peewee==3.18.3
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
protobuf==6.33.1
psycopg2-binary==2.9.11
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydash==8.0.5
Pygments==2.19.2
pyparsing==3.2.5
pytest==9.0.1
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.5
sec-api==1.0.32
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.66
